requirements.txt
fastapi
uvicorn[standard]
python-jose
passlib[bcrypt]
sqlalchemy
psycopg2-binary
